First version of the sorting functionality of the pgnparser.
This version acknowledges sorting criteria with an arbitrary number of
keys and directions (either increasing or decreasing). There is
however a problem: tags are stored as strings and therefore
comparisons are made as if they were string and hence "5" goes after
"49".

A better version should take care of this and to sort variables taking
into account their type.