This is the first version attempting at providing the functionality of automatically generating histograms.
The major change is that histograms of any order (implemented in
pgnhistogram.go) are implemented. In spite of the name of the file,
these are generic histograms that could be used in any
package. Additionally, a unit test is provided to ensure that
histograms work as expected.

Additionally, pgnparser already recognizes the flag "histogram" and it
even provides help on it. However, histograms are not implemented yet
in the core of pgnparser. Only small modifications have been done to
pgncollection and pgntools