Package table added
Years ago, pgnparser came with its own implementation of tables in a
nested package named "tbl". Actually, the development of the pgnparser
was interrupted to create an independente package for displaying tables
which is already available in github.

This changeset substitutes the representation of tables with the new
package "table"