fstools removed
With the availability of a PgnFile which handles all the burden to
access files and process games, the fstools are not needed anymore.

In addition, this changeset fixes a bug in the outcome of some weird
games. If no output was recorded then the PGN format writes '*' and this
was not considered in the grouped regexp processing the outcome. The fix
consists of considering a special outcome {-1, -1} which means that no
result was given