First version of the metatemplate package
The metatemplate package allows the usage of metavariables in templates.
Meta-variables appear in the form ${variable} and they are substituted
by their value which can be found in a map. Additionally, the user can
provide a default value with ${variable[default:<default-value>]} and/or
a prompt in the form ${variable[prompt:<user prompt>]}. In case a prompt
is given, the user is prompted for a value and in case a default value
was provided also with ${value[prompt:<user
prompt>][default:default-value]}, the user is provided the default value
which is accepted if RET is pressed immediately. In case no prompt is
given, the default value is used if any is given.

Meta-templating allows the usage of the same template with many
different purposes