.tex generation for lichess pgn files
Lichess pgn files do not provide a comment at the end of the game. This
comment was used as a token when producing the .tex output for fics
games. A slight modification has been done to produce correct output for
lichess pgn games as well in the absence of a last comment