Meta-templates almost behave as regular templates
Still, Execute is ran over an instance of a template. All in all, this
version of metatemplate.go includes various modifications which are all
based on the definition of an explicit type, MetaTemplate. The idea is
to use it as a regular template with extended functionality for a few
services ---only those required by pgnparser